+++ src/download/index.ts
@@ -15,6 +15,7 @@ import { isBinaryInstalled } from '../external.js';
 import type { BrowserCookie } from '../types.js';
 import { getErrorMessage } from '../errors.js';
 import { fetchWithNodeNetwork } from '../node-network.js';
+import { log } from '../logger.js';
 
 export type { BrowserCookie } from '../types.js';
 
@@ -267,7 +268,7 @@ export async function ytdlpDownload(
       if (fs.existsSync(cookiesFile)) {
         args.push('--cookies', cookiesFile);
       } else {
-        console.error(`[download] Cookies file not found: ${cookiesFile}, falling back to browser cookies`);
+        log.warn(`[download] Cookies file not found: ${cookiesFile}, falling back to browser cookies`);
         args.push('--cookies-from-browser', 'chrome');
       }
     } else {
